refactor: move coingeckoTop500 to auxLists and split into multiple files
