+++ package.json
@@ -10,7 +10,7 @@
     "coingecko": "yarn run-script src/scripts/coingecko.js",
     "downloadImages": "yarn run-script src/scripts/downloadImages",
     "cowFi:tokens": "yarn run-script ./src/scripts/cowFi-tokens.js",
-    "coingeckoTop500": "node src/scripts/coingeckoTop500.js",
+    "generateAuxLists": "node src/scripts/auxLists",
     "uniswapToGnosisChainList": "yarn run-script src/scripts/mapTokenListToBridge.ts 100 uniswap",
     "coingeckoToGnosisChainList": "yarn run-script src/scripts/mapTokenListToBridge.ts 100 coingecko",
     "coingeckoToArbitrumOneList": "yarn run-script src/scripts/mapTokenListToBridge.ts 42161 coingecko",
